Fix bug with destination path: The path in GetDatasetFileByVersion response is actually an S3 key (datasetId/path).
