+++ rehydrate/fargate/go.mod
@@ -8,7 +8,6 @@ require (
 	github.com/aws/aws-sdk-go v1.45.23
 	github.com/aws/aws-sdk-go-v2 v1.24.1
 	github.com/aws/aws-sdk-go-v2/service/s3 v1.48.1
-	github.com/google/uuid v1.3.0
 	github.com/pennsieve/pennsieve-go v1.3.1
 	github.com/pennsieve/rehydration-service/shared v0.0.0-00010101000000-000000000000
 	github.com/stretchr/testify v1.8.4
@@ -41,6 +40,7 @@ require (
 	github.com/aws/smithy-go v1.19.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/golang-jwt/jwt v3.2.2+incompatible // indirect
+	github.com/google/uuid v1.6.0 // indirect
 	github.com/jmespath/go-jmespath v0.4.0 // indirect
 	github.com/pennsieve/pennsieve-go-api v1.1.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
