+++ rehydrate/shared/go.mod
@@ -14,6 +14,7 @@ require (
 	github.com/aws/aws-sdk-go-v2/service/s3 v1.48.1
 	github.com/aws/aws-sdk-go-v2/service/sqs v1.29.7
 	github.com/aws/smithy-go v1.19.0
+	github.com/google/uuid v1.6.0
 	github.com/pennsieve/pennsieve-go v1.3.1
 	github.com/stretchr/testify v1.8.4
 )
